SEM05/requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
# requirements.txt — Dependințe Python pentru Seminar 05
# ═══════════════════════════════════════════════════════════════════════════════
# Sisteme de Operare | ASE București - CSIE
# Seminar 05: Advanced Bash Scripting
#
# INSTALARE:
#   pip install -r requirements.txt
#   sau
#   pip install --user -r requirements.txt
#
# PENTRU DEZVOLTARE:
#   pip install -r requirements.txt -r requirements-dev.txt
# ═══════════════════════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────────────────
# DEPENDINȚE PRINCIPALE (necesare pentru funcționalitate de bază)
# ─────────────────────────────────────────────────────────────────────────────

# Parser YAML - necesar pentru quiz.yaml
pyyaml>=6.0

# Terminal formatting - pentru output colorat în quiz_runner.py
rich>=13.0

# ─────────────────────────────────────────────────────────────────────────────
# DEPENDINȚE OPȚIONALE (pentru funcționalități extinse)
# ─────────────────────────────────────────────────────────────────────────────

# Validare JSON Schema (opțional, pentru quiz_lms.json)
# jsonschema>=4.0

# Progress bars (opțional, pentru procesări lungi)
# tqdm>=4.0

# ─────────────────────────────────────────────────────────────────────────────
# DEPENDINȚE DEZVOLTARE (pentru contribuitori)
# ─────────────────────────────────────────────────────────────────────────────
# Acestea sunt opționale și folosite doar pentru dezvoltare/testare

# Linting Python
# ruff>=0.1.0

# Type checking
# mypy>=1.0

# Testing
# pytest>=7.0
# pytest-cov>=4.0

# ─────────────────────────────────────────────────────────────────────────────
# NOTE
# ─────────────────────────────────────────────────────────────────────────────
# 
# Versiuni minime testate:
#   - Python 3.9+
#   - PyYAML 6.0+
#   - Rich 13.0+
#
# Pe Ubuntu/Debian, instalează și:
#   sudo apt install shellcheck
#
# Pe WSL2, asigură-te că ai Python3:
#   sudo apt install python3 python3-pip
#
# ═══════════════════════════════════════════════════════════════════════════════
